move parse_filter to synapse.api.filtering
This is useful for deduplication. I am using this
on my websocket-implementation so I try do deduplicate this
part. But it might be useful for other sync approaches as well.
