+++ changelog.d/2393.misc
@@ -1,0 +1,1 @@
+move parse_filter to synapse.api.filtering
