+++ synapse/rest/client/v2_alpha/_base.py
@@ -51,16 +51,6 @@ def client_patterns(path_regex, releases=(0,), unstable=True, v1=False):
     return patterns
 
 
-def set_timeline_upper_limit(filter_json, filter_timeline_limit):
-    if filter_timeline_limit < 0:
-        return  # no upper limits
-    timeline = filter_json.get("room", {}).get("timeline", {})
-    if "limit" in timeline:
-        filter_json["room"]["timeline"]["limit"] = min(
-            filter_json["room"]["timeline"]["limit"], filter_timeline_limit
-        )
-
-
 def interactive_auth_handler(orig):
     """Wraps an on_POST method to handle InteractiveAuthIncompleteErrors
